requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
black==25.9.0
cachetools==6.2.1
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distlib==0.4.0
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
flake8==7.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.49.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
identify==2.6.15
idna==3.11
ImageHash==4.3.2
iniconfig==2.3.0
isort==7.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.4
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
pytokens==0.2.0
PyWavelets==1.9.0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scipy==1.16.3
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.35.3
websockets==15.0.1
pillow-heif